start a schematic for an external preamp
wanting a separate module as an overdrive unit. 
This amp seems to be designed for clean tone, full bass response. Overdriving the tubes in the amp as it is wired just does not sound very good at all.
Inspired by "tube driver" guitar pedal, and a preamp circuit from ax86.com